Cleaning and moving props part 18
