--- src/main/java/com/paneedah/mwc/models/weapons/ImpaledBody.java
@@ -1,103 +1,0 @@
-// Date: 2/17/2019 10:08:49 PM
-// Template version 1.1
-// Java generated by Techne
-// Keep in mind that you still need to fill in some blanks
-// - ZeuX
-
-
-
-
-
-
-package com.paneedah.mwc.models.weapons;
-
-import net.minecraft.client.model.ModelBase;
-import net.minecraft.client.model.ModelRenderer;
-import net.minecraft.entity.Entity;
-
-public class ImpaledBody extends ModelBase
-{
-  //fields
-    ModelRenderer head;
-    ModelRenderer body;
-    ModelRenderer rightarm;
-    ModelRenderer leftarm;
-    ModelRenderer rightleg;
-    ModelRenderer leftleg;
-    ModelRenderer rebar;
-  
-  public ImpaledBody()
-  {
-    textureWidth = 128;
-    textureHeight = 128;
-    
-      head = new ModelRenderer(this, 0, 0);
-      head.addBox(-4F, -8F, -4F, 8, 8, 8);
-      head.setRotationPoint(0F, -9F, 0F);
-      head.setTextureSize(128, 128);
-      head.mirror = true;
-      setRotation(head, -0.9666439F, 0F, 0F);
-      body = new ModelRenderer(this, 16, 16);
-      body.addBox(-4F, 0F, -2F, 8, 12, 4);
-      body.setRotationPoint(0F, -9F, 0F);
-      body.setTextureSize(128, 128);
-      body.mirror = true;
-      setRotation(body, -0.4461433F, 0F, 0F);
-      rightarm = new ModelRenderer(this, 40, 16);
-      rightarm.addBox(-3F, -2F, -2F, 4, 12, 4);
-      rightarm.setRotationPoint(-5F, -7F, -0.5F);
-      rightarm.setTextureSize(128, 128);
-      rightarm.mirror = true;
-      setRotation(rightarm, 0F, 0F, 0.1115358F);
-      leftarm = new ModelRenderer(this, 40, 16);
-      leftarm.addBox(-1F, -2F, -2F, 4, 12, 4);
-      leftarm.setRotationPoint(5F, -7F, -0.5F);
-      leftarm.setTextureSize(128, 128);
-      leftarm.mirror = true;
-      setRotation(leftarm, 0.1858931F, 0F, 0F);
-      rightleg = new ModelRenderer(this, 0, 16);
-      rightleg.addBox(-2F, 0F, -2F, 4, 12, 4);
-      rightleg.setRotationPoint(-2F, 1.5F, -5F);
-      rightleg.setTextureSize(128, 128);
-      rightleg.mirror = true;
-      setRotation(rightleg, 0F, 0F, 0.1487144F);
-      leftleg = new ModelRenderer(this, 0, 16);
-      leftleg.addBox(-2F, 0F, -2F, 4, 12, 4);
-      leftleg.setRotationPoint(2F, 1.5F, -5F);
-      leftleg.setTextureSize(128, 128);
-      leftleg.mirror = true;
-      setRotation(leftleg, 0.0743572F, 0F, 0F);
-      rebar = new ModelRenderer(this, 0, 36);
-      rebar.addBox(-4F, 0F, -2F, 2, 50, 2);
-      rebar.setRotationPoint(3F, -25F, -6F);
-      rebar.setTextureSize(128, 128);
-      rebar.mirror = true;
-      setRotation(rebar, 0.1858931F, 0F, 0F);
-  }
-  
-  public void render(Entity entity, float f, float f1, float f2, float f3, float f4, float f5)
-  {
-    super.render(entity, f, f1, f2, f3, f4, f5);
-    setRotationAngles(f, f1, f2, f3, f4, f5, entity);
-    head.render(f5);
-    body.render(f5);
-    rightarm.render(f5);
-    leftarm.render(f5);
-    rightleg.render(f5);
-    leftleg.render(f5);
-    rebar.render(f5);
-  }
-  
-  private void setRotation(ModelRenderer model, float x, float y, float z)
-  {
-    model.rotateAngleX = x;
-    model.rotateAngleY = y;
-    model.rotateAngleZ = z;
-  }
-  
-  public void setRotationAngles(float f, float f1, float f2, float f3, float f4, float f5, Entity entity)
-  {
-    super.setRotationAngles(f, f1, f2, f3, f4, f5, entity);
-  }
-
-}
